+++ CHANGELOG.md
@@ -16,9 +16,12 @@
 ### Additions and Improvements
  - Added beacon-api `/eth/v1/beacon/states/{state_id}/pending_partial_withdrawals` endpoint for use post-electra.
  - Added beacon-api `/eth/v1/beacon/states/{state_id}/pending_deposits` endpoint for use post-electra.
+ - Added beacon-api `/eth/v1/beacon/states/{state_id}/pending_consolidations` endpoint for use post-electra.
  - Added Chiado Electra configuration due at Mar-06-2025 09:43:40 GMT+0000
  - Removed stack trace by default from duty failure messages.
  - Holesky pectra bad block ignored to aid syncing
+ - Added a development flag to increase the maximum pending queue for attestations.
+ - Added support for [Hoodi testnet](https://github.com/eth-clients/hoodi/).
 
 ### Bug Fixes
  - Added 415 response code for beacon-api `/eth/v1/validator/register_validator`.